refactor(examples): extract fetch module
